An explicit yes is what permits ordinary correspondence
The business_correspondence arm of VerdictForPerson read qualifying events
and never the person's own recorded answer, so a `granted` person_consent
row authorized nothing. A person who had said in as many words that we may
write to them was refused until they happened to send an inbound message,
and one inbound message then opened every unrelated send.

correspondenceVerdict reads the recorded state first: an explicit grant
allows on its own. The implied arm is unchanged and stays second, because it
is what answers for the overwhelming majority of people, who record no
answer either way. objectionStands already runs before the class switch, so
a withdrawal still beats both — pinned rather than assumed.

purpose.RequiresDOI is now consulted for this class. It is false for
correspondence today; hard-coding that here would have silently ignored an
installation that turned it on.

The large-file ceiling forced a split: the qualifying-event readers move to
qualifyingground.go unchanged, which puts the recorded answer and the
implied ground in separate files in the order they are asked.

Found by driving real mail through the real send doors on a dev stack, not
by a failing test — no test in the tree covered a grant on this class.
