+++ backend/internal/modules/consent/verdict.go
@@ -132,25 +132,7 @@ func VerdictForPerson(ctx context.Context, tx pgx.Tx, personID string, purpose P
 		return Verdict{State: VerdictAllowed, Reason: "account and contract notices need no consent"}, nil
 
 	case ClassBusinessCorrespondence:
-		event, source, err := latestQualifyingEvent(ctx, tx, personID)
-		if err != nil {
-			return Verdict{}, err
-		}
-		if source == sourceNone {
-			// No inbound, no inquiry, no deal, no recorded exchange. There is
-			// nothing here to balance, so this is not the easy Art 6(1)(f) case
-			// and the honest answer is that nobody has decided.
-			return Verdict{
-				State:  VerdictUnknown,
-				Reason: "they have never written to you and no deal or inquiry connects you",
-			}, nil
-		}
-		return Verdict{
-			State:             VerdictAllowed,
-			Reason:            qualifyingReason(event),
-			Qualifying:        &event,
-			QualifyingDerived: source == sourceDerived,
-		}, nil
+		return correspondenceVerdict(ctx, tx, personID, purpose)
 
 	case ClassPhoneOutreach:
 		// Dormant by decision: the purpose exists so the model is complete. A
@@ -160,6 +142,66 @@ func VerdictForPerson(ctx context.Context, tx pgx.Tx, personID string, purpose P
 	default:
 		return marketingVerdict(ctx, tx, personID, purpose)
 	}
+}
+
+// correspondenceVerdict answers for ordinary business correspondence: the
+// person's own recorded answer first, and only then what the record implies.
+//
+// THE ORDER IS THE FIX. This arm used to read qualifying events alone, so an
+// explicit `granted` row authorized nothing: a person who had said in as many
+// words that we may write to them was refused until they happened to send an
+// inbound message, while one inbound message opened every unrelated send. A
+// recorded answer is the strongest thing a subject can give us about their own
+// correspondence, and reading it second — or, as here, not at all — inverted
+// that.
+//
+// A withdrawal needs no branch here. objectionStands runs before this for every
+// class and every purpose, so a withdrawn row has already blocked.
+//
+// The implied arm below is unchanged and stays second: it is the Art 6(1)(f)
+// reading of a relationship the subject started, and it is what answers for the
+// overwhelming majority of people, who never record an answer either way.
+func correspondenceVerdict(ctx context.Context, tx pgx.Tx, personID string, purpose PurposeRow) (Verdict, error) {
+	// requiresDOI is passed as the purpose declares it rather than as a
+	// constant false. Correspondence does not demand the round trip today, and
+	// hard-coding that here would silently ignore an installation that turned
+	// it on for this purpose.
+	state, granted, err := recordedState(ctx, tx, personID, purpose.ID, purpose.RequiresDOI)
+	if err != nil {
+		return Verdict{}, err
+	}
+	if granted {
+		return Verdict{State: VerdictAllowed, Reason: "they said we may write to them"}, nil
+	}
+	if state == string(StateGranted) && purpose.RequiresDOI {
+		// Granted but never confirmed, on an installation that demands the
+		// round trip for this purpose. Falling through to the implied arm
+		// would let the timeline supply what the confirmation did not.
+		return Verdict{
+			State:  VerdictBlocked,
+			Reason: "consent was recorded but never confirmed by the double opt-in",
+			Code:   BlockUnconfirmedDOI,
+		}, nil
+	}
+	event, source, err := latestQualifyingEvent(ctx, tx, personID)
+	if err != nil {
+		return Verdict{}, err
+	}
+	if source == sourceNone {
+		// No answer on file, no inbound, no inquiry, no deal, no recorded
+		// exchange. There is nothing here to balance, so this is not the easy
+		// Art 6(1)(f) case and the honest answer is that nobody has decided.
+		return Verdict{
+			State:  VerdictUnknown,
+			Reason: "they have never written to you and no deal or inquiry connects you",
+		}, nil
+	}
+	return Verdict{
+		State:             VerdictAllowed,
+		Reason:            qualifyingReason(event),
+		Qualifying:        &event,
+		QualifyingDerived: source == sourceDerived,
+	}, nil
 }
 
 // marketingVerdict is the strict arm, unchanged in strictness by ADR-0098:
@@ -261,152 +303,6 @@ func objectionStands(ctx context.Context, tx pgx.Tx, personID, purposeID string)
 		return false, time.Time{}, fmt.Errorf("read the objection's proof: %w", err)
 	}
 	return true, at, nil
-}
-
-// latestQualifyingEvent finds the most recent thing on the record that makes
-// correspondence lawful.
-//
-// It looks in two places, and the order is the point. A TYPED row wins: an
-// in-person exchange or an inquiry is something a named human recorded, and it
-// carries the note or the reference a dispute asks for. Failing that the
-// captured timeline answers for itself — an inbound message from this person IS
-// the qualifying event, which is what "deterministic and derivable from
-// captured data" means (ADR-0098 D2).
-//
-// Deriving rather than requiring a written row is what keeps the rule honest on
-// day one: every mailbox this product has ever captured already contains the
-// evidence, and a model that only recognised events recorded after this build
-// shipped would tell a rep they may not answer somebody who wrote to them last
-// week.
-func latestQualifyingEvent(ctx context.Context, tx pgx.Tx, personID string) (QualifyingEvent, eventSource, error) {
-	event, found, err := recordedQualifyingEvent(ctx, tx, personID)
-	if err != nil {
-		return QualifyingEvent{}, sourceNone, err
-	}
-	if found {
-		return event, sourceRecorded, nil
-	}
-	event, found, err = inboundQualifyingEvent(ctx, tx, personID)
-	if err != nil {
-		return QualifyingEvent{}, sourceNone, err
-	}
-	if found {
-		return event, sourceDerived, nil
-	}
-	return QualifyingEvent{}, sourceNone, nil
-}
-
-// eventSource says where a qualifying event came from, and therefore whether
-// the transmit path still owes it a stamp.
-//
-// One value rather than two adjacent bools: `found` and `derived` are the same
-// type in neighbouring positions, so nothing but care stopped a future arm
-// returning them the wrong way round — and returning them the wrong way round
-// means either stamping a duplicate of a row that already exists, or relying on
-// a basis that was never written down.
-type eventSource int
-
-const (
-	// sourceNone: nothing on the record makes correspondence lawful.
-	sourceNone eventSource = iota
-	// sourceRecorded: a stored row already proves it, so there is nothing to stamp.
-	sourceRecorded
-	// sourceDerived: read off the timeline, and the transmit path must record it
-	// before relying on it (Art 5(2)).
-	sourceDerived
-)
-
-// RecordDerivedQualifyingEvent stamps a derived qualifying event onto the
-// record, so what authorized a send is a fact somebody can look up rather than
-// a computation this build happened to make.
-//
-// ADR-0098 D2 requires the flip be "stamped with which event and when", and
-// Art 5(2) is why: a lawful basis nobody wrote down is an assertion, and the
-// controller carries the burden of showing it. Deriving the event at read time
-// answers the question correctly; it does not answer it accountably.
-//
-// Only the TRANSMIT path calls this. A preview authorizes nothing, and writing
-// a legal fact because somebody opened a composer would record a basis for a
-// message that was never sent.
-//
-// The insert is idempotent on the source record — the same inbound message
-// re-derived on the next send must not stack a second row claiming a second
-// event happened — and the guarantee is the database's unique index, not a
-// check this function performs and a concurrent caller races past.
-func RecordDerivedQualifyingEvent(ctx context.Context, tx pgx.Tx, personID string, event QualifyingEvent, capturedBy string) error {
-	// ON CONFLICT, not NOT EXISTS: two concurrent sends to the same person both
-	// pass a read-then-write check and both insert. The unique index on the
-	// source record is what actually makes this idempotent.
-	_, err := tx.Exec(ctx, `
-		INSERT INTO consent_qualifying_event
-			(person_id, kind, source_entity_type, source_entity_id,
-			 occurred_at, source, captured_by)
-		VALUES ($1, $2, $3, $4, $5, 'derived', $6)
-		ON CONFLICT (person_id, source_entity_type, source_entity_id)
-		  WHERE source_entity_id IS NOT NULL
-		  DO NOTHING`,
-		personID, event.Kind, event.SourceEntityType, event.SourceEntityID,
-		event.OccurredAt, capturedBy)
-	if err != nil {
-		return fmt.Errorf("consent: stamp the qualifying event that allowed this send: %w", err)
-	}
-	return nil
-}
-
-// recordedQualifyingEvent reads a row a human or an integration wrote.
-func recordedQualifyingEvent(ctx context.Context, tx pgx.Tx, personID string) (QualifyingEvent, bool, error) {
-	var event QualifyingEvent
-	var sourceType, sourceID, note *string
-	err := tx.QueryRow(ctx, `
-		SELECT kind, occurred_at, source_entity_type, source_entity_id, note
-		FROM consent_qualifying_event
-		WHERE person_id = $1
-		ORDER BY occurred_at DESC
-		LIMIT 1`, personID).Scan(&event.Kind, &event.OccurredAt, &sourceType, &sourceID, &note)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return QualifyingEvent{}, false, nil
-	}
-	if err != nil {
-		return QualifyingEvent{}, false, fmt.Errorf("read the qualifying event: %w", err)
-	}
-	if sourceType != nil {
-		event.SourceEntityType = *sourceType
-	}
-	if sourceID != nil {
-		event.SourceEntityID = *sourceID
-	}
-	if note != nil {
-		event.Note = *note
-	}
-	return event, true, nil
-}
-
-// inboundQualifyingEvent derives the event from the captured timeline: they
-// wrote to us, and the message itself is the proof.
-//
-// The activity is reached through activity_link, the same table every
-// person-scoped timeline read walks, so this cannot count a message the record
-// does not show.
-func inboundQualifyingEvent(ctx context.Context, tx pgx.Tx, personID string) (QualifyingEvent, bool, error) {
-	var event QualifyingEvent
-	var activityID string
-	err := tx.QueryRow(ctx, `
-		SELECT a.id, a.occurred_at
-		FROM activity a
-		JOIN activity_link l ON l.activity_id = a.id AND l.person_id = $1
-		WHERE a.direction = 'inbound' AND a.archived_at IS NULL
-		ORDER BY a.occurred_at DESC
-		LIMIT 1`, personID).Scan(&activityID, &event.OccurredAt)
-	if errors.Is(err, pgx.ErrNoRows) {
-		return QualifyingEvent{}, false, nil
-	}
-	if err != nil {
-		return QualifyingEvent{}, false, fmt.Errorf("read the inbound qualifying message: %w", err)
-	}
-	event.Kind = "inbound_message"
-	event.SourceEntityType = "activity"
-	event.SourceEntityID = activityID
-	return event, true, nil
 }
 
 // recordedState reads the person's own decision for this purpose, and whether
